fix all bugs e implementado token e autorização bearer

+++ services/user.service.js
@@ -2,8 +2,8 @@ const db = require('../db')
 var md5 = require('md5')
 const service = require('./login.service')
 
-module.exports.getAllUsers = async (obj) => {
-    token_validate = await service.validate(obj)
+module.exports.getAllUsers = async (token) => {
+    token_validate = await service.validate(token.split(' ')[1])
     if(token_validate.success == true){
         [record] = await db.query("SELECT * FROM users")
         data = { "success": true, "data": record }
@@ -13,11 +13,10 @@ module.exports.getAllUsers = async (obj) => {
     return data;
 }
 
-module.exports.getUserById = async (obj, id) => {
-    token_validate = await service.validate(obj)
+module.exports.getUserById = async (token, id) => {
+    token_validate = await service.validate(token.split(' ')[1])
     if(token_validate.success == true){
-        [[record]] = await db.query("SELECT * FROM users WHERE id = ?", [id])
-        console.log(record)
+        [record] = await db.query("SELECT * FROM users WHERE id = ?", [id])
         data = { "success": true, "data": record }
     } else {
         data = { success: false, error: "Token Inválido" };
@@ -25,28 +24,53 @@ module.exports.getUserById = async (obj, id) => {
     return data;
 }
 
-module.exports.deleteUser = async (obj, id) => {
-    token_validate = await service.validate(obj)
-    const [{ affectedRows }] = await db.query("DELETE FROM users WHERE id = ?", [id])
-    return affectedRows;
-}
-
-module.exports.addUser = async (obj) => {
-    token_validate = await service.validate(obj)
-    const [{ affectedRows }] = await db.query("INSERT INTO users (name, user, password) VALUES (?, ?, ?)", 
-        [obj.name, obj.user, md5(obj.password)])
-    return affectedRows;
-}
-
-module.exports.updateUser = async (obj, id) => {
-    token_validate = await service.validate(obj)
-    const [records] = await db.query("SELECT password FROM users WHERE id = ?", [id])
-    if (records[0].password == md5(obj.password)) {
-        password = records[0].password
+module.exports.deleteUser = async (token, id) => {
+    token_validate = await service.validate(token.split(' ')[1])
+    if(token_validate.success == true){
+        const [{ affectedRows }] = await db.query("DELETE FROM users WHERE id = ?", [id])
+        if(affectedRows == 1){
+            data = { "success": true, "data": "Usuário apagado" }
+        } else {
+            data = { "success": true, "data": "Usuário já deletado" }
+        }
     } else {
-        password = md5(obj.password)
+        data = { success: false, error: "Token Inválido" };
     }
-    const [{ affectedRows }] = await db.query("UPDATE users SET name = ?, user = ?, password = ? WHERE id = ?", 
-        [obj.name, obj.user, password, id])
-    return affectedRows;
+    return data;
+}
+
+module.exports.addUser = async (token, obj) => {
+    token_validate = await service.validate(token.split(' ')[1])
+    if(token_validate.success == true){
+        const [{ affectedRows }] = await db.query("INSERT INTO users (name, user, password) VALUES (?, ?, ?)", [obj.name, obj.user, md5(obj.password)])
+        if(affectedRows == 1){
+            data = { "success": true, "data": "Usuário criado com sucesso" }
+        } else {
+            data = { "success": true, "data": "Houve um erro ao cadastrar o usuário" }
+        }
+    } else {
+        data = { success: false, error: "Token Inválido" };
+    }
+    return data;
+}
+
+module.exports.updateUser = async (token, id, obj) => {
+    token_validate = await service.validate(token.split(' ')[1])
+    if(token_validate.success == true){
+        const [records] = await db.query("SELECT password FROM users WHERE id = ?", [id])
+        if (records[0].password == md5(obj.password)) {
+            password = records[0].password
+        } else {
+            password = md5(obj.password)
+        }
+        const [{ affectedRows }] = await db.query("UPDATE users SET name = ?, user = ?, password = ? WHERE id = ?",  [obj.name, obj.user, password, id])
+        if(affectedRows == 1){
+            data = { "success": true, "data": "Usuário Atualizado com sucesso" }
+        } else {
+            data = { "success": true, "data": "Houve um erro ao atualizar o usuário" }
+        }
+    } else {
+        data = { success: false, error: "Token Inválido" };
+    }
+    return data;
 }
